created validation for all http methods

+++ src/main/java/io/github/cursospring/libraryapi/service/AutorService.java
@@ -1,7 +1,10 @@
 package io.github.cursospring.libraryapi.service;
 
+import io.github.cursospring.libraryapi.exceptions.OperacaoNaoPermitidaException;
 import io.github.cursospring.libraryapi.model.Autor;
 import io.github.cursospring.libraryapi.repository.AutorRepository;
+import io.github.cursospring.libraryapi.repository.LivroRepository;
+import io.github.cursospring.libraryapi.validator.AutorValidator;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -12,20 +15,25 @@ import java.util.UUID;
 public class AutorService {
 
     private final AutorRepository repository;
+    private final AutorValidator validator;
+    private final LivroRepository livroRepository;
 
-    public AutorService(AutorRepository repository) {
+    public AutorService(AutorRepository repository, AutorValidator validator, LivroRepository livroRepository) {
+        this.validator = validator;
         this.repository = repository;
+        this.livroRepository = livroRepository;
     }
 
     public Autor salvar(Autor autor){
+        validator.validar(autor);
         return repository.save(autor);
     }
 
     public void atualizar(Autor autor){
         if (autor.getId() == null){
             throw new IllegalArgumentException("Para atualizar é necessario que o autor esteja salvo na base.");
-
         }
+        validator.validar(autor);
         repository.save(autor);
     }
 
@@ -34,6 +42,9 @@ public class AutorService {
     }
 
     public void deletar (Autor autor){
+        if(possuiLivro(autor)){
+            throw new OperacaoNaoPermitidaException("Não é permitido excluir autor que possui livros cadastrados.");
+        }
         repository.delete(autor);
     }
 
@@ -53,4 +64,7 @@ public class AutorService {
         return repository.findAll();
     }
 
+    public boolean possuiLivro(Autor autor){
+        return livroRepository.existsByAutor(autor);
+    }
 }
